refactor: unify module naming — kaoyan→graduateExam, secondhand→marketplace
- Rename core/kaoyan/ + core/kaoyanquery/ → core/graduateExam/
  Postgraduate* classes → GraduateExam*
- Rename core/secondhand/ → core/marketplace/
  Secondhand* classes → Marketplace*
- Rename frontend views/kaoyan/ → views/graduateExam/
- Rename frontend views/ershou/ → views/marketplace/
- Rename frontend api/kaoyan.js → api/graduateExam.js
- Update DataSource config, router, and external references
- API URL paths unchanged (/api/kaoyan/query, /api/ershou/*)
